one more pass at the dbg, intern, internp modules
