+++ macros/src/function_like/intern.rs
@@ -3,7 +3,7 @@ use syn::{parse_macro_input, LitStr};
 
 use crate::construct;
 
-pub(crate) fn expand(input: TokenStream) -> TokenStream {
-    let lit = parse_macro_input!(input as LitStr);
-    construct::interned_string(&lit.value(), "str", false).into()
+pub(crate) fn expand(args: TokenStream) -> TokenStream {
+    let literal = parse_macro_input!(args as LitStr);
+    construct::interned_string(&literal.value(), "str", false).into()
 }
