+++ macros/src/function_like/internp.rs
@@ -4,15 +4,15 @@ use syn::{parse_macro_input, LitStr};
 
 use crate::construct;
 
-pub(crate) fn expand(input: TokenStream) -> TokenStream {
-    let lit = parse_macro_input!(input as LitStr);
-    let sym_name = construct::mangled_symbol_name("prim", &lit.value());
+pub(crate) fn expand(args: TokenStream) -> TokenStream {
+    let literal = parse_macro_input!(args as LitStr);
+    let sym_name = construct::mangled_symbol_name("prim", &literal.value());
 
     let prefix = Some("prim");
     let section = construct::linker_section(false, prefix, &sym_name);
     let section_for_macos = construct::linker_section(true, prefix, &sym_name);
 
-    let sym = if cfg!(feature = "unstable-test") {
+    let var_addr = if cfg!(feature = "unstable-test") {
         quote!({ defmt::export::fetch_add_string_index() as u16 })
     } else {
         quote!({
@@ -25,7 +25,7 @@ pub(crate) fn expand(input: TokenStream) -> TokenStream {
     };
 
     quote!({
-        defmt::export::make_istr(#sym)
+        defmt::export::make_istr(#var_addr)
     })
     .into()
 }
